+++ chord_metadata_service/restapi/models.py
@@ -1,7 +1,5 @@
-from chord_metadata_service.chord.models import Project
+from abc import abstractmethod
 from django.db import models
-from abc import ABC, abstractproperty, abstractmethod
-from enum import Enum
 
 
 class IndexableMixin:
@@ -11,26 +9,13 @@ class IndexableMixin:
 
 
 class SchemaType(models.TextChoices):
-    PHENOPACKET = "PHENOPACKET"
-    BIOSAMPLE = "BIOSAMPLE"
-    INDIVIDUAL = "INDIVIDUAL"
-
-
-class ExtraSchema(models.Model):
-    id = models.CharField(primary_key=True, max_length=200)
-    project = models.ForeignKey(Project, on_delete=models.CASCADE, related_name="project_schemas")
-    extra_schema = models.JSONField()
-    schema_type = models.CharField(max_length=200, choices=SchemaType)
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(fields=["project", "schema_type"], name="unique_project_schema")
-        ]
+    PHENOPACKET = ("PHENOPACKET", "PHENOPACKET")
+    BIOSAMPLE = ("BIOSAMPLE", "BIOSAMPLE")
+    INDIVIDUAL = ("INDIVIDUAL", "INDIVIDUAL")
 
 
 class BaseExtraProperties(models.Model):
     extra_properties = models.JSONField()
-    extra_schema = models.ForeignKey(ExtraSchema, on_delete=models.CASCADE, related_name="")
 
     class Meta:
         abstract = True
