+++ pyproject.toml
@@ -57,7 +57,6 @@ tabulate = "^0.9.0"
 uritemplate = "^4.1.1"
 
 [tool.poetry.group.dev.dependencies]
-codecov = "^2.1.12"
 coverage = "^6.5.0"
 flake8 = "^6.0.0"
 tox = "^4.2.8"
